FIX "db.Model" is not defined error
Applying typing for flask_sqlalchemy has been a trouble as we will
need a correct mypy plugin for the models. Since `Model` is monkey
patched to the db object, mypy does not know that exist.

Declaring db.Model as `BaseModel` is a workaround for now, but we
will add a stub for this in the future. The candidate for this
will be https://github.com/ssfdust/flask-sqlalchemy-stubs.