FIX add necessary None-checks
In various places, we access attributes of objects that are possibly
None, which might lead to a runtime error.

In some of the cases, we know that the object is logically non-None.
We need to add assertions to make this assumption explicit to the
type-checker.

In other cases, we need to manually check if the object is None
and decide what to do.