CHG refactor `get_request_file_stream` for verbose typing
With `typing.overload` we can infer the type from the first
argument of this function. It gives us a better sense of
how the return type of this will look like when invoking.

A behavioral change is made to ensure that `secure_filename`
does not get called with `None`, which violates the type
annotation of it.