ADD an option to force push via sidecar API.
This allows the student to force push via the CLI. Making `git push` and
`git push -f` available should be sufficient for students to maneuver with
Git on our IDE instances for the purposes of learning. It's unlikely
that we are going to support more arguments.

Allowing the students to push via Anubis CLI is only for the convenience
of not requiring them to set up their own authentication mechanism on
GitHub or storing their SSH keys or GitHub token on the server.

Signed-off-by: Zixuan James Li <p359101898@gmail.com>